feat: send kafka message when connector is reset
